Created a way to save state and continue upon restarting CookForMe
The GetFromDatabase function will fetch the user's previous state from a mysql database stored one Amazon Webservices RTS.
When a user says the command "continue" at Cook For Me's start, the lambda function will make a query to the database to see what the previous state was. If the user previously was in the middle of a recipe the the GetInstructionStepByStep method will be called and will start at the step where it left off.